+++ .vscode/settings.json
@@ -108,7 +108,13 @@
         "inttypes.h": "c",
         "ssd1306.h": "c",
         "splashscreen.h": "c",
-        "assert.h": "c"
+        "assert.h": "c",
+        "esp_gatt_common_api.h": "c",
+        "stdint.h": "c",
+        "mqtt_client.h": "c",
+        "esp_log.h": "c",
+        "timer.h": "c",
+        "display.h": "c"
     },
 	"files.exclude": {
 		"**/.git": true,
