+++ main/CMakeLists.txt
@@ -1,4 +1,4 @@
-set(COMPONENT_SRCS "blemqttproxy.c")
+set(COMPONENT_SRCS "blemqttproxy.c helperfunctions.c")
 set(COMPONENT_ADD_INCLUDEDIRS ".")
 
 register_component()
