+++ main/beacon.c
@@ -1,0 +1,112 @@
+
+
+
+uint16_t s_active_beacon_mask = 0;
+
+
+
+uint8_t beacon_maj_min_to_idx(uint16_t maj, uint16_t min)
+{
+    if( (maj == CONFIG_BLE_DEVICE_1_MAJ ) && (min == CONFIG_BLE_DEVICE_1_MIN ) ) return 0;
+    if( (maj == CONFIG_BLE_DEVICE_2_MAJ ) && (min == CONFIG_BLE_DEVICE_2_MIN ) ) return 1;
+    if( (maj == CONFIG_BLE_DEVICE_3_MAJ ) && (min == CONFIG_BLE_DEVICE_3_MIN ) ) return 2;
+    if( (maj == CONFIG_BLE_DEVICE_4_MAJ ) && (min == CONFIG_BLE_DEVICE_4_MIN ) ) return 3;
+    if( (maj == CONFIG_BLE_DEVICE_5_MAJ ) && (min == CONFIG_BLE_DEVICE_5_MIN ) ) return 4;
+    if( (maj == CONFIG_BLE_DEVICE_6_MAJ ) && (min == CONFIG_BLE_DEVICE_6_MIN ) ) return 5;
+    if( (maj == CONFIG_BLE_DEVICE_7_MAJ ) && (min == CONFIG_BLE_DEVICE_7_MIN ) ) return 6;
+    if( (maj == CONFIG_BLE_DEVICE_8_MAJ ) && (min == CONFIG_BLE_DEVICE_8_MIN ) ) return 7;
+    if( (maj == CONFIG_BLE_DEVICE_9_MAJ ) && (min == CONFIG_BLE_DEVICE_9_MIN ) ) return 8;
+    if( (maj == CONFIG_BLE_DEVICE_10_MAJ) && (min == CONFIG_BLE_DEVICE_10_MIN) ) return 9;
+
+    ESP_LOGE(TAG, "beacon_maj_min_to_idx: unknown maj %d min %d", maj, min);
+
+    return UNKNOWN_BEACON;
+}
+
+__attribute__((unused)) uint8_t num_active_beacon()
+{
+    uint8_t num_act_beac = 0;
+    for(int i=0; i < CONFIG_BLE_DEVICE_COUNT_USE; i++){
+        if(is_beacon_idx_active(i)){
+            num_act_beac++;
+        }
+    }
+    return num_act_beac;
+}
+
+__attribute__((unused)) uint8_t first_active_beacon()
+{
+    uint8_t first_act_beac = UNKNOWN_BEACON;
+
+    for(int i=0; i < CONFIG_BLE_DEVICE_COUNT_USE; i++){
+        if(is_beacon_idx_active(i)){
+            first_act_beac = i;
+        }
+    }
+
+    return first_act_beac;
+}
+
+bool is_beacon_idx_active(uint16_t idx)
+{
+    return (s_active_beacon_mask & (1 << idx) );
+}
+
+void set_beacon_idx_active(uint16_t idx)
+{
+    s_active_beacon_mask |= (1 << idx);
+    persist_active_beacon_mask();
+}
+
+__attribute__((unused)) void clear_beacon_idx_active(uint16_t idx)
+{
+    s_active_beacon_mask &= ~(1 << idx);
+    persist_active_beacon_mask();
+}
+
+__attribute__((unused)) bool toggle_beacon_idx_active(uint16_t idx)
+{
+    s_active_beacon_mask ^= (1 << idx);
+    persist_active_beacon_mask();
+
+    return s_active_beacon_mask & (1 << idx);
+}
+
+void persist_active_beacon_mask()
+{
+    esp_err_t err;
+    UNUSED(err);
+
+    if(CONFIG_ACTIVE_BLE_DEVICE_PERSISTANCE){
+        err = save_blemqttproxy_param();
+    }
+}
+
+beacon_type_t esp_ble_is_mybeacon_packet (uint8_t *adv_data, uint8_t adv_data_len, uint8_t scan_rsp_len){
+    beacon_type_t result = UNKNOWN_BEACON;
+
+    // ESP_LOG_BUFFER_HEXDUMP(TAG, adv_data, adv_data_len, ESP_LOG_DEBUG);
+
+    if ((adv_data != NULL) && (adv_data_len == 0x1E)){
+        if ( (!memcmp(adv_data, (uint8_t*)&mybeacon_common_head_v3, sizeof(mybeacon_common_head_v3)))
+            && (!memcmp((uint8_t*)(adv_data + sizeof(mybeacon_common_head_v3)),
+                (uint8_t*)&mybeacon_common_vendor_v3, sizeof(mybeacon_common_vendor_v3.proximity_uuid)))){
+            ESP_LOGD(TAG, "esp_ble_is_mybeacon_packet v3, true");
+            result = BEACON_V3;
+        } else {
+            ESP_LOGD(TAG, "esp_ble_is_mybeacon_packet v3, false");
+            result = UNKNOWN_BEACON;
+        }
+    } else if ((adv_data != NULL) && (adv_data_len == 0x17)){
+        if ( (!memcmp(adv_data, (uint8_t*)&mybeacon_common_head_v4, sizeof(mybeacon_common_head_v4))) ){
+            ESP_LOGD(TAG, "esp_ble_is_mybeacon_packet v4, true");
+
+            result = (scan_rsp_len ? BEACON_V4_SR : BEACON_V4);
+        } else {
+            ESP_LOGD(TAG, "esp_ble_is_mybeacon_packet v4, false");
+            result = UNKNOWN_BEACON;
+        }
+    }
+
+    return result;
+}
